Added some multiprocessing
Converting large images to SIXEL with the default options is now a lot faster on multi-core systems. On single core systems (or on systems where ProcessPoolExecutor isn't available), the performance will likely to go down a bit.

WARNING: Register counts over 256 are now not supported at all and will likely raise an error (they weren't supposed to be allowed in the first place).

TODO:
* Make it faster
* Cleanup the code
* Error handling